Change EdgeVertexIndex to usize, remove all "as usize"

+++ map-to-3D/src/edge.rs
@@ -9,7 +9,7 @@ use std::fmt;
 use std::str::FromStr;
 use std::num::ParseIntError;
 
-pub type EdgeVertexIndex = i32;
+pub type EdgeVertexIndex = usize;
 
 /// An "Edge" - a connection between two points
 #[derive(Hash, PartialEq, Eq, Ord, PartialOrd, Clone, Debug, Copy)]
